scrapper/requirements.txt
selenium==4.21.0
requests==2.32.3
numpy==2.0.0
pillow==10.3.0